añadido log y mejorado legibilidad y control de errores
+Se ha añadido un log
+Se ha mejorado los controladores
+Se han controlado las mayoría de errores
+Se ha mejorado el orden en la librería de Servicios
+Se ha mejorado la creacion de Servicios con una factoria

+++ Repository/Repositories/GameRepository.cs
@@ -1,7 +1,10 @@
 ﻿using Core.Models;
 using Domain.Interfaces;
-using Repository.Data;
 using Microsoft.EntityFrameworkCore;
+using Repository.Data;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace Repository.Repositories
 {
@@ -16,45 +19,55 @@ namespace Repository.Repositories
 
         public async Task<List<Game>> GetAllAsync()
         {
-            return await _context.Games.ToListAsync();
+            return await _context.Games.ToListAsync().ConfigureAwait(false);
         }
 
         public async Task<Game> GetByIdAsync(int id)
         {
-            return await _context.Games.FindAsync(id);
+            return await _context.Games.FindAsync(id).ConfigureAwait(false);
         }
 
         public async Task<Game> CreateAsync(Game game)
         {
             _context.Games.Add(game);
-            await _context.SaveChangesAsync();
+            await _context.SaveChangesAsync().ConfigureAwait(false);
             return game;
         }
 
         public async Task<bool> UpdateAsync(Game game)
         {
             _context.Entry(game).State = EntityState.Modified;
-            return await _context.SaveChangesAsync() > 0;
+            return await _context.SaveChangesAsync().ConfigureAwait(false) > 0;
         }
 
         public async Task<bool> DeleteAsync(int id)
         {
-            var game = await _context.Games.FindAsync(id);
+            var gameExists = await _context.Games.AnyAsync(g => g.Id == id).ConfigureAwait(false);
+            if (!gameExists)
+                return false;
+
+            var game = await _context.Games.FindAsync(id).ConfigureAwait(false);
             if (game == null)
                 return false;
 
             _context.Games.Remove(game);
-            return await _context.SaveChangesAsync() > 0;
+            return await _context.SaveChangesAsync().ConfigureAwait(false) > 0;
         }
 
         public async Task<List<Game>> GetGamesByPlatformAsync(string platform)
         {
-            return await _context.Games.Where(g => g.Platform == platform).ToListAsync();
+            return await _context.Games
+                .Where(g => g.Platform == platform)
+                .ToListAsync()
+                .ConfigureAwait(false);
         }
 
         public async Task<List<Game>> GetGamesByGenreAsync(string genre)
         {
-            return await _context.Games.Where(g => g.Genre == genre).ToListAsync();
+            return await _context.Games
+                .Where(g => g.Genre == genre)
+                .ToListAsync()
+                .ConfigureAwait(false);
         }
     }
 }
